+++ restaurant-platform-api/src/app.module.ts
@@ -14,6 +14,7 @@ import { CategoriesModule } from './modules/categories/categories.module';
 import { MenuItemsModule } from './modules/menu-items/menu-items.module';
 import { OrdersModule } from './modules/orders/orders.module';
 import { FeedbackModule } from './modules/feedback/feedback.module';
+import { WebsocketsModule } from './websockets/websockets.module';
 import { HealthController } from './health.controller';
 import { GlobalExceptionFilter } from './common/filters/http-exception.filter';
 import { TransformInterceptor } from './common/interceptors/transform.interceptor';
@@ -49,6 +50,7 @@ import { JwtAuthGuard } from './common/guards';
     MenuItemsModule,
     OrdersModule,
     FeedbackModule,
+    WebsocketsModule,
   ],
   controllers: [HealthController],
   providers: [