+++ restaurant-platform-api/src/modules/orders/orders.module.ts
@@ -2,8 +2,10 @@ import { Module } from '@nestjs/common';
 import { OrdersService } from './orders.service';
 import { OrdersController } from './orders.controller';
 import { OrdersCustomerController } from './orders-customer.controller';
+import { WebsocketsModule } from '../../websockets/websockets.module';
 
 @Module({
+  imports: [WebsocketsModule],
   controllers: [OrdersController, OrdersCustomerController],
   providers: [OrdersService],
   exports: [OrdersService],